Switch the license back to BSD.
The 'ros2' branch of robot_state_publisher is clearly derived
from the ROS 1 port (that lives on the 'indigo-devel' and
'melodic-devel' branches).  Thus, the license should be BSD.
There were a few community contributions to the 'ros2' branch
under the (erroneous) Apache 2.0 license, but all contributors
have signified their willingness to license back to BSD in
https://github.com/ros/robot_state_publisher/issues/120 .
This change switches the license back to the correct BSD
license on the 'ros2' branch.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>