Make sure not to crash on an invalid URDF.
That is, once we are up and running, we only accept changes
to the URDF that successfully parse.  If the URDF doesn't parse,
return an error to the caller, print an error to the console,
and continue using the old one.

Also fix a small bug where we forgot to set the robot_description
parameter when we are using the fallback command-line strategy.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>