Change how parameter updates are handled.
set_parameter callbacks should not have any side-effects.
This is most easy to understand if you think about the fact
that those callbacks can be chained; if callback A has a side-effect,
but then callback B rejects the entire set_parameter call, then
the nodes will be in an inconsistent state.

Instead, change it so that the set_parameter callback checks that
the incoming arguments are valid, and does nothing else.  We
setup a callback to listen to the /parameter_events topic;
at that point, we know that the update has been accepted by
all set_parameter callbacks, so it is safe for side-effects to happen.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>